+++ src/ModelingToolkit.jl
@@ -117,7 +117,7 @@ include("utils.jl")
 include("domains.jl")
 
 # Code that should eventually go elsewhere, but is here for fow
-include("compat/bareiss.jl")
+include("systems/bareiss.jl")
 if !isdefined(Graphs, :IncrementalCycleTracker)
     include("compat/incremental_cycles.jl")
 end
